ref : 유저, 센터 더미 데이터 생성 클래스 분리
